requirements.txt
aes==1.2.0
altair==4.2.2
altgraph==0.17.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Crypto.Cipher==1
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
des==1.0.6
distlib==0.3.5
Django==4.1.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.0.0
fastjsonschema==2.16.1
ffmpeg==1.4
ffmpeg-downloader==0.2.0
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fonttools==4.37.1
frozendict==2.3.5
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httplib2==0.21.0
idna==3.3
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.14.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
keyboard==0.13.5
kiwisolver==1.4.4
lib50==3.0.4
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.1
mistune==2.0.4
MouseInfo==0.1.3
moviepy==1.0.3
multitasking==0.0.11
mysql-connector==2.2.9
mysql-connector-python==8.0.32
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
oauth2client==4.1.3
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.13.0
prettytable==3.6.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.0
PyDrive==1.3.1
pydub==0.25.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
pythonautogui==0.0.2
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.7
PyYAML==5.4.1
pyzmq==23.2.1
requests==2.28.1
rich==13.3.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.12.1
selenium==4.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
stack-data==0.5.0
streamlit==1.20.0
submit50==3.1.1
tabulate==0.9.0
telebot==0.0.5
tenacity==8.2.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
ticker==0.13
tinycss2==1.1.1
tinytag==1.8.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
utils==1.0.1
validators==0.20.0
virtualenv==20.16.3
watchdog==2.3.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yfinance==0.2.12
zipp==3.15.0
